Add a tag-saving service
Just for now - to be able to save the tags, so that we can teach the
maschine :). Later would done differently.
I don't want the tags to be duplicated so I've used the _id as well (as
well because reading it I shouldn't care - I should just take a name)